chore [DS-239] wip // cleaned starknet helper
