+++ src/actions/blockchain/utils.ts
@@ -1,6 +1,6 @@
 import { action, z } from "@daydreamsai/core";
-import { formatTokenBalance, normalizeAddress, getTokenBalance } from "../../utils/starknet";
-import { executeQuery, getGameSessionId } from "../../utils/graphql";
+import { formatTokenBalance, getTokenBalance } from "../../utils/starknet";
+import { executeQuery, getGameSessionId, normalizeAddress } from "../../utils/graphql";
 import { GET_AGENT_LIQUIDITY_POSITIONS, GET_AGENT_FARM_POSITIONS } from "../../utils/queries";
 import {
   getAgentAddress,
