+++ src/utils/starknet.ts
@@ -1,15 +1,28 @@
+/**
+ * Starknet Utility Functions
+ * This file contains utility functions for interacting with the Starknet blockchain,
+ * including configuration management, value conversion, and common contract operations.
+ */
+
 import { StarknetChain } from "@daydreamsai/defai";
 import { uint256, cairo, type Call } from "starknet";
 import { StarknetConfigStore } from "../agents/utils";
 
-// Create a function to get the current agent ID
-// This will be set by the agent when it starts
+// Global variable to track the current agent's ID
 let currentAgentId: string | null = null;
 
+/**
+ * Sets the current agent ID for Starknet operations
+ * @param agentId - The ID of the agent to set as current
+ */
 export function setCurrentAgentId(agentId: string) {
   currentAgentId = agentId;
 }
 
+/**
+ * Gets the current agent ID, returns a default if none is set
+ * @returns The current agent ID or 'default-agent' as fallback
+ */
 export function getCurrentAgentId(): string {
   if (!currentAgentId) {
     console.warn(
@@ -20,7 +33,11 @@ export function getCurrentAgentId(): string {
   return currentAgentId;
 }
 
-// Function to get Starknet configuration for the current agent
+/**
+ * Retrieves Starknet configuration for the current agent
+ * First tries to get agent-specific config, then falls back to environment variables
+ * @returns Configuration object with RPC URL and credentials
+ */
 const getStarknetConfig = () => {
   // Try to get agent-specific configuration first
   if (currentAgentId) {
@@ -69,7 +86,10 @@ const getStarknetConfig = () => {
   };
 };
 
-// Create a function to get a Starknet chain instance for the current agent
+/**
+ * Creates and returns a StarknetChain instance for the current agent
+ * @returns Configured StarknetChain instance
+ */
 export const getStarknetChain = () => {
   const config = getStarknetConfig();
   return new StarknetChain({
@@ -79,35 +99,13 @@ export const getStarknetChain = () => {
   });
 };
 
-// For backward compatibility - lazy initialization to avoid errors at module load time
-let _starknetChain: StarknetChain | null = null;
-export const starknetChain = {
-  read: async (params: any) => {
-    if (!_starknetChain) {
-      _starknetChain = getStarknetChain();
-    }
-    return _starknetChain.read(params);
-  },
-  write: async (params: any) => {
-    if (!_starknetChain) {
-      _starknetChain = getStarknetChain();
-    }
-    return _starknetChain.write(params);
-  },
-  executeMulticall: async (calls: Call[]) => {
-    if (!_starknetChain) {
-      _starknetChain = getStarknetChain();
-    }
-    return _starknetChain.writeMulticall(calls);
-  },
-};
-
-// Contract value conversion utilities
-export const convertToContractValue = (amount: string, decimals: number): string => {
-  return BigInt(Math.floor(Number(amount) * 10 ** decimals)).toString();
-};
-
-// Token approval utility
+/**
+ * Creates an ERC20 token approval call
+ * @param tokenAddress - Address of the token contract
+ * @param spenderAddress - Address of the spender
+ * @param amount - Amount to approve
+ * @returns Call object for token approval
+ */
 export const getApproveCall = (
   tokenAddress: string,
   spenderAddress: string,
@@ -121,7 +119,11 @@ export const getApproveCall = (
   };
 };
 
-// Execute multiple calls in a single transaction
+/**
+ * Executes multiple contract calls in a single transaction
+ * @param calls - Array of Call objects to execute
+ * @returns Transaction result
+ */
 export const executeMultiCall = async (calls: Call[]) => {
   try {
     // Execute multicall
@@ -134,13 +136,22 @@ export const executeMultiCall = async (calls: Call[]) => {
   }
 };
 
-// Helper function to convert to uint256
+/**
+ * Converts a value to Cairo uint256 format and returns low and high parts
+ * @param value - Value to convert
+ * @returns Tuple of [low, high] parts of uint256
+ */
 export const toUint256WithSpread = (value: string): [string, string] => {
   const uint = cairo.uint256(value);
   return [uint.low.toString(), uint.high.toString()];
 };
 
-// Helper function to convert u256 to decimal
+/**
+ * Converts a Cairo uint256 to decimal format
+ * @param low - Low part of uint256
+ * @param high - High part of uint256
+ * @returns Decimal value as BigInt
+ */
 export const convertU256ToDecimal = (low: string, high: string) => {
   try {
     const u256Value = uint256.uint256ToBN({
@@ -155,10 +166,10 @@ export const convertU256ToDecimal = (low: string, high: string) => {
 };
 
 /**
- * Convert a BigInt value to hexadecimal string format
- * @param value BigInt value or string representing a BigInt
- * @param withPrefix Whether to include '0x' prefix (default: true)
- * @returns Hexadecimal string representation
+ * Converts a BigInt value to hexadecimal format
+ * @param value - BigInt value or string to convert
+ * @param withPrefix - Whether to include '0x' prefix
+ * @returns Hexadecimal string
  */
 export const toHex = (value: bigint | string, withPrefix: boolean = true): string => {
   try {
@@ -172,9 +183,10 @@ export const toHex = (value: bigint | string, withPrefix: boolean = true): strin
 };
 
 /**
- * Calculate optimal liquidity amounts using Uniswap's constant product formula
- * @param params Parameters for calculating optimal liquidity
- * @returns Optimal amounts for both tokens and whether it's first provision
+ * Calculates optimal liquidity amounts for Uniswap-style pools
+ * Uses constant product formula (x * y = k)
+ * @param params - Pool parameters including token addresses and amounts
+ * @returns Optimal amounts for both tokens and first provision status
  */
 export const calculateOptimalLiquidity = async (params: {
   contractAddress: string;
@@ -261,22 +273,11 @@ export const calculateOptimalLiquidity = async (params: {
 };
 
 /**
- * Removes trailing zeros from a Starknet address and ensures proper hex format
- * @param address The address to normalize
- * @returns Normalized address without trailing zeros
+ * Retrieves token balance for a given address
+ * @param contractAddress - Token contract address
+ * @param playerAddress - Address to check balance for
+ * @returns Balance as BigInt
  */
-export const normalizeAddress = (address: string): string => {
-  // Ensure the address starts with '0x'
-  const prefixedAddress = address.startsWith("0x") ? address : `0x${address}`;
-
-  // Remove trailing zeros and maintain '0x' prefix
-  const normalized = prefixedAddress.toLowerCase().replace(/^0x0*/, "0x");
-
-  // If the result is just '0x', return '0x0'
-  return normalized === "0x" ? "0x0" : normalized;
-};
-
-// Helper function to get balance and convert to decimal
 export const getTokenBalance = async (contractAddress: string, playerAddress: string) => {
   const chain = getStarknetChain();
   const result = await chain.read({
@@ -287,7 +288,22 @@ export const getTokenBalance = async (contractAddress: string, playerAddress: st
   return convertU256ToDecimal(result[0], result[1]);
 };
 
-// Helper function to format token balance
+/**
+ * Formats a raw token balance to human-readable format
+ * Assumes 18 decimal places
+ * @param rawBalance - Raw balance as BigInt
+ * @returns Formatted balance string
+ */
 export const formatTokenBalance = (rawBalance: bigint) => {
   return (Number(rawBalance) / Math.pow(10, 18)).toString();
+};
+
+/**
+ * Converts a decimal amount to contract value with specified decimals
+ * @param amount - Amount as a decimal string
+ * @param decimals - Number of decimal places
+ * @returns Contract value as string
+ */
+export const convertToContractValue = (amount: string, decimals: number): string => {
+  return BigInt(Math.floor(Number(amount) * 10 ** decimals)).toString();
 };
